feat(Dialog plugin): Ability to access Router, Vuex store etc within custom component #4009
