+++ docs/src/pages/quasar-plugins/dialog.md
@@ -68,8 +68,13 @@ import CustomComponent from '..path.to.component..'
 this.$q.dialog({
   component: CustomComponent,
 
+  // optional if you want to have access to
+  // Router, Vuex store, and so on, in your
+  // custom component:
+  root: this.$root,
+
   // props forwarded to component
-  // (everything except "component" prop above):
+  // (everything except "component" and "root" props above):
   text: 'something',
   // ...more.props...
 }).onOk(() => {
